feat!: separate subscribe methods from request methods
BREAKING CHANGE: You now need to call `listen()` for subscription methods,
and `request()` for request methods. Previously, we use a single `request`
method and checked whether the given method was a "subscribe" method or not,
and called the underlying correct path (eg either `method` directly,
or `method_subscribe`).

We've got a PR at https://github.com/sanity-io/bifur/pull/45 where in my mind it
would make sense to reuse the same name for a method for both a subscription and
a request, eg `authorization` (performs authorization), `authorization_subscribe`
(listens for changes to the authorization). This wouldn't have been possible in
the client without picking a different method name. While that would be an option,
I actually think it is somewhat cleaner to specify "what you want" - this way
the reader of the consuming code will not be in doubt about whether a call is a
one-of request method or one that will provide updates.